requirements.txt
absl-py==0.15.0
antlr4-python3-runtime==4.8
astunparse==1.6.3
avro-python3==1.10.2
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
dpcpp-cpp-rt==2022.0.3
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.25.0
fvcore==0.1.5.post20220414
gast==0.4.0
gdown==4.4.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.11.2
intel-cmplr-lib-rt==2022.0.3
intel-cmplr-lic-rt==2022.0.3
intel-opencl-rt==2022.0.3
intel-openmp==2022.0.3
iopath==0.1.9
itsdangerous==2.1.1
Jinja2==3.0.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
mkl==2022.0.3
mkl-fft==1.3.0
mkl-service==2.4.0
munkres==1.1.4
numpy==1.19.2
oauthlib==3.2.0
omegaconf==2.1.2
opencv-contrib-python==4.1.2.30
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
ply==3.11
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==302
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
sip==6.6.1
six==1.15.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tbb==2021.5.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf-slim==1.1.0
toml==0.10.2
torch==1.8.2+cpu
torchaudio==0.11.0
torchvision==0.9.2+cpu
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.12.1
yacs==0.1.8
zipp==3.7.0
